WIP weird enum issue for importing the js string across projects. Solution is to make main = the types declaration file but that doesn't feel right
