+++ packages/generator-particle/package.json
@@ -20,6 +20,7 @@
     "yeoman-generator": "^4.11.0"
   },
   "devDependencies": {
+    "@phase2/particle-cli": "file:../particle-cli",
     "@types/yeoman-generator": "^3.1.4",
     "typescript": "^3.9.6",
     "yo": "^3.1.1"
