+++ packages/particle-cli/src/generatePromptOptions.ts
@@ -9,6 +9,7 @@ import {
   ComponentLibraryOptions,
   FrontendFrameworkOptions,
   TestingLibraryOptions,
+  // @ts-ignore
 } from './types'
 
 const prompt = inquirer.createPromptModule()
