+++ packages/particle-cli/test/create.test.ts
@@ -1,8 +1,14 @@
+// import mock from 'mock-fs'
+// @ts-ignore
+import mockSpawn from 'mock-spawn'
+
 import create from '../src/create'
 import repoPackage from '../package.json'
 
-const { name } = repoPackage
+const spawn = mockSpawn()
+require('child_process').spawn = spawn
 
+const { name } = repoPackage
 describe(`${name}/create`, () => {
   it('show log', () => {
     expect(create({})).toBeTruthy()
